requirements.txt
numpy
pandas
Flask
scikit-learn
requests
tensorflow
Pillow
gunicorn == 20.0.4